* bugfixing: incorrect round during calculations * delete as button instead of FAB (in progress - done for debt and trip)

+++ app/src/main/java/com/macgavrina/co_accounting/interfaces/AddReceiverInAddDebtContract.kt
@@ -32,7 +32,7 @@ interface AddReceiverInAddDebtContract {
 
     interface Presenter:BasePresenterContract<View> {
 
-        fun amountIsEdited(newAmount: Float)
+        fun amountIsEdited(newAmount: Double)
 
         fun cancelButtonInToolbarIsClicked()
